Disable 'Clean Data Directory' per game
Prerequisite for Enderal, otherwise it crashes on launch after using WB
with it.

Background: Enderal ships an empty folder called 'JCData', which
JContainers requires in order to function - if it's removed, JContainers
will crash the entire game. Normally, this folder contains some dummy
files and example Lua scripts, which is why it never gets removed for
standard JContainers installations.

But Enderal, as already mentioned, ships it empty - JContainers is
fine with this, but WB's 'Clean Data Directory' feature, which is
enabled by default, will remove the folder, causing Enderal to crash.

Things this commit does:
1. Rename it to 'Remove Empty Directories', because that's what it
   does.
2. Add a boolean that can force-disable it for a game and will also
   hide the toggle for it.

TODO - flip the default to False and add a warning when it's flipped?
Or even remove it altogether? This could cause issues with all kinds of
mods that purposefully ship empty directories and expect them to exist.