cosaves.py - beta
This API was starting to become inadequate for Wrye Bash's current
needs. Most importantly:
 1. It was difficult to understand and maintain.
 2. It did not support enough of the cosave format.
 3. Pluggy support was all over the place.

This commit looks like a complete rewrite, but it was actually a
step-by-step refactoring comprising 100+ commits that got squashed down
to this final commit, since almost every step of the way would have
broken dev.

For reference:
  cosaves.py - alpha: 0a300af01a85bb3535d3194203fd34cc0d3e9b29.

New features included in this commit are:

1. Add button for dumping Pluggy cosaves

  With the new cosaves API, we now have the capability to dump Pluggy
  cosaves as well, so this commit adds a button to do just that.

2. Correctly display save masters in SSE

  With SKSE 2.0.15, which adds the PLGN chunk, we can now display save
  master lists in SSE in the correct order (with ESLs and regular plugins
  interweaved).