Switch gitPython with pygit2
This should take care of the previous errors with gitPython either not
finding git or it being on a non standard path. pygit2 comes bundled
with libgit2, meaning no need to find or have git installed (one less
dependency :D).

The API is a bit awkward but since we don't do much with it it shouldn't
really matter.