Show race tweaks only for Oblivion
These seem to be wholly Oblivion-specific, there isn't much point in
showing them for the other games.

As the TODO mentions, this is a pretty disgusting solution. I tried
ripping them into a new game-specific file (similar to
game/*/patcher/special.py), but that turned into a mess due to how bush
imports stuff from there. So this seems to be the least intrusive
solution for now.