Esmify/Espify Masters Fixups
A while ago, I renamed 'Esmify Self' to 'Add ESM Flag' and 'Espify Self'
to 'Remove ESM Flag'. That commit also contained an edit to 'Espify
Masters', but that edit was incorrect - it shows 'Add ESM Flag to
Masters' when it should be showing 'Remove ESM Flag From Masters'.

I took this opportunity to make all the 'Esmify/Espify' ocurrences
consistent and to finally edit the Readme for these changes.